INFRA-758 merge and enhance puppet-swagger-generator
This change merges the puppet-swagger-generator code and allows to
generate directly into the code.
It also now supports openapi 2.0 api server definitions instead
of the swagger 1.2 (which are deprecated).
Instead of parsing all known models, the generator now only parses
the models that can be created in a kubernetes cluster (there's no
point managing a kubernetes object that can't be created in the
cluster).