INFRA-764 don't force match between puppet resource and k8s
The problem with the previous way of creating k8s object from puppet
resources was that the puppet resource name was used as the name
of the k8s resource. This was problematic because the puppet name is
global to a catalog (ie cluster), but not the k8s name (which is global
in a namespace).
Thus we had to decorate all puppet resource name with the namespace
in a way that the k8s object is doubly namespaced (ie api-api in the
for api namespace instance).

To solve this problem, this change introduces a few new things:
* if the puppet resource name looks like namespace::name, then the
module will try to look up resources with that scheme. The object
metadata must of course match. This allows to solve the issue of the
puppet global namespace. We'll be able to have a puppet resource
kubernetes_deployment { 'api::api': }, while the k8s object will be
called 'api' in the namespace 'api'.

* at the same time, we store in every k8s object created an annotations
that contains the puppet resource name. This means that the puppet
name can now be completely unrelated to the name used in k8s (as long
as the object is defined with metadata).

* when looking for puppet resources to match with the known k8s object,
we do look by FQDN, the annotations, and the simple name. This should
normally cover all cases including legacy puppet resource names until
we change them all.